separate setTriangle
