+++ src/MeshBVH.js
@@ -4,13 +4,13 @@ import { CENTER } from './Constants.js';
 import { buildTree } from './buildFunctions.js';
 import { OrientedBox } from './Utils/OrientedBox.js';
 import { SeparatingAxisTriangle } from './Utils/SeparatingAxisTriangle.js';
+import { setTriangle } from './Utils/TriangleUtils.js';
 import { sphereIntersectTriangle } from './Utils/MathUtilities.js';
 import {
 	raycast,
 	raycastFirst,
 	shapecast,
 	intersectsGeometry,
-	setTriangle,
 } from './castFunctions.js';
 
 import {
